feat: role based auth

+++ controllers/user.controller.js
@@ -5,7 +5,9 @@ import {
   deleteUserByEmail 
 } from '../services/index.js'
 
-export const createUserController = async (req, res) => {
+import { isAdmin } from '../middleware/auth.js'
+
+export const createUserController = async (req, res, next) => {
   try {
     const user = await createUser(req.body.email, req.body.password)
     res.status(201).send({
@@ -15,9 +17,10 @@ export const createUserController = async (req, res) => {
   } catch (error) {
     res.status(500).send('Error al registrar el usuario')
   }
+  next()
 }
 
-export const findUserByEmailController = async (req, res) => {
+export const findUserByEmailController = async (req, res, next) => {
   try {
     const user = await findUserByEmail(req.body.email)
     res.status(200).send({
@@ -27,9 +30,10 @@ export const findUserByEmailController = async (req, res) => {
   } catch (error) {
     res.status(500).send('Error al encontrar el usuario')
   }
+  next()
 }
 
-export const updateUserByEmailController = async (req, res) => {
+export const updateUserByEmailController = async (req, res, next) => {
   try {
     const user = await updateUserByEmail(req.body.email, req.body.data)
     res.status(200).send({
@@ -39,16 +43,22 @@ export const updateUserByEmailController = async (req, res) => {
   } catch (error) {
     res.status(500).send('Error al actualizar el usuario')
   }
+  next()
 }
 
-export const deleteUserByEmailController = async (req, res) => {
+export const deleteUserByEmailController = async (req, res, next) => {
   try {
-    const user = await deleteUserByEmail(req.body.email)
-    res.status(200).send({
-      ...user,
-      message: 'Usuario borrado correctamente'
-    })
+    if (isAdmin(req.user)) {
+      const user = await deleteUserByEmail(req.body.email)
+      res.status(200).send({
+        ...user,
+        message: 'Usuario borrado correctamente'
+      })
+    }else{
+      res.status(400).send("Error you need to be an admin to do that");
+    }
   } catch (error) {
     res.status(500).send('Error al borrar el usuario')
   }
+  next()
 }
